chore: move to correct packaging and remove unnecessary code
