+++ src/main/java/com/epam/spring/service/impl/TraineeService.java
@@ -19,8 +19,8 @@ import com.epam.spring.model.User;
 import com.epam.spring.repository.TokenRepository;
 import com.epam.spring.repository.TraineeRepository;
 import com.epam.spring.repository.TrainerRepository;
-import com.epam.spring.service.JwtService;
-import com.epam.spring.service.MyUserPrincipal;
+import com.epam.spring.service.auth.JwtService;
+import com.epam.spring.service.auth.MyUserPrincipal;
 import com.epam.spring.service.base.TraineeSpecificOperationsService;
 import com.epam.spring.util.PasswordGenerator;
 import com.epam.spring.util.TransactionContext;
