+++ src/main/java/com/epam/spring/service/impl/TrainerService.java
@@ -16,8 +16,8 @@ import com.epam.spring.repository.TokenRepository;
 import com.epam.spring.repository.TraineeRepository;
 import com.epam.spring.repository.TrainerRepository;
 import com.epam.spring.repository.TrainingTypeRepository;
-import com.epam.spring.service.JwtService;
-import com.epam.spring.service.MyUserPrincipal;
+import com.epam.spring.service.auth.JwtService;
+import com.epam.spring.service.auth.MyUserPrincipal;
 import com.epam.spring.service.base.TrainerSpecificOperationsService;
 import com.epam.spring.util.PasswordGenerator;
 import com.epam.spring.util.TransactionContext;
